+++ src/extension.ts
@@ -20,15 +20,17 @@ import {
   USER_KITS_FILEPATH,
   findCLCompilerPath,
   effectiveKitEnvironment,
+  scanForKitsIfNeeded,
 } from '@cmt/kit';
 import {KitsController} from '@cmt/kitsController';
 import * as logging from '@cmt/logging';
 import {fs} from '@cmt/pr';
 import {FireNow, FireLate} from '@cmt/prop';
 import rollbar from '@cmt/rollbar';
+import {StateManager} from './state';
 import {StatusBar} from '@cmt/status';
 import * as telemetry from '@cmt/telemetry';
-import {ProjectOutlineProvider, TargetNode, SourceFileNode} from '@cmt/tree';
+import {ProjectOutlineProvider, TargetNode, SourceFileNode, WorkspaceFolderNode} from '@cmt/tree';
 import * as util from '@cmt/util';
 import {ProgressHandle, DummyDisposable, reportProgress} from '@cmt/util';
 import {DEFAULT_VARIANTS} from '@cmt/variant';
@@ -58,7 +60,7 @@ type CMakeToolsQueryMapFn = (cmt: CMakeTools) => Thenable<string | string[] | nu
 class ExtensionManager implements vscode.Disposable {
   constructor(public readonly extensionContext: vscode.ExtensionContext) {
     telemetry.activate();
-    this._statusBar.targetName = 'all';
+    this._statusBar.setBuildTargetName('all');
     this._folders.onAfterAddFolder(async cmtFolder => {
       console.assert(this._folders.size === vscode.workspace.workspaceFolders?.length);
       if (this._folders.size === 1) {
@@ -125,6 +127,21 @@ class ExtensionManager implements vscode.Disposable {
       }
       this._statusBar.setAutoSelectActiveFolder(v);
     });
+
+    this._workspaceConfig.onChange('sourceDirectory', async v => {
+      await this._onDidChangeActiveTextEditorSub.dispose();
+      if (v) {
+        this._onDidChangeActiveTextEditorSub = vscode.window.onDidChangeActiveTextEditor(e => this._onDidChangeActiveTextEditor(e), this);
+      } else {
+        this._onDidChangeActiveTextEditorSub = new DummyDisposable();
+      }
+
+      // Reset to a full activation mode.
+      // If required, the extension will switch back to partial activation mode
+      // after more analysis triggered by the below configure.
+      await enableFullFeatureSet(true);
+      await vscode.commands.executeCommand("cmake.configure");
+    });
   }
 
   private _onDidChangeActiveTextEditorSub: vscode.Disposable = new DummyDisposable();
@@ -165,6 +182,32 @@ class ExtensionManager implements vscode.Disposable {
     telemetry.logEvent('open', telemetryProperties);
   }
 
+  public getActiveFolderContext() : StateManager | undefined {
+    const activeFolder = this.getActiveFolder();
+    if (activeFolder) {
+      const extensionState = new StateManager(this.extensionContext, activeFolder);
+      return extensionState;
+    }
+
+    return undefined;
+  }
+
+  // Partial activation means that the CMake Tools commands are hidden
+  // from the commands pallette and the status bar is not visible.
+  // The context variable "cmake:enableFullFeatureSet" is always equal
+  // to the state setting ignoreCMakeListsMissing.
+  // To have them both always in sync, cmake:enableFullFeatureSet is set
+  // by the getter and setter of ignoreCMakeListsMissing.
+  public enableFullFeatureSet(fullFeatureSet: boolean) {
+    const context = this.getActiveFolderContext();
+    if (context) {
+      context.ignoreCMakeListsMissing = !fullFeatureSet;
+      this._statusBar.setVisible(fullFeatureSet);
+    } else {
+      log.trace(localize('enableFullFeatureSet.no.active.folder', 'enableFullFeatureSet({0}) called but not active folder found.'), fullFeatureSet);
+    }
+  }
+
   /**
    * Create a new extension manager instance. There must only be one!
    * @param ctx The extension context
@@ -179,11 +222,18 @@ class ExtensionManager implements vscode.Disposable {
    * The folder controller manages multiple instances. One per folder.
    */
   private readonly _folders = new CMakeToolsFolderController(this.extensionContext);
+  public getActiveFolder(): vscode.WorkspaceFolder | undefined {
+    if (!vscode.workspace.workspaceFolders || vscode.workspace.workspaceFolders.length === 0) {
+      return undefined;
+    }
+
+    return this._folders?.activeFolder?.cmakeTools.folder || vscode.workspace.workspaceFolders![0];
+  }
 
   /**
    * The status bar controller
    */
-  private readonly _statusBar = new StatusBar();
+  private readonly _statusBar = new StatusBar(this._workspaceConfig);
   // Subscriptions for status bar items:
   private _statusMessageSub: vscode.Disposable = new DummyDisposable();
   private _targetNameSub: vscode.Disposable = new DummyDisposable();
@@ -233,10 +283,11 @@ class ExtensionManager implements vscode.Disposable {
       // Expected schema is file...
       return vscode.workspace.getWorkspaceFolder(vscode.Uri.file(folder as string));
     }
-    if ((folder as vscode.WorkspaceFolder).uri) {
-      return folder;
+    const workspaceFolder = folder as vscode.WorkspaceFolder;
+    if (util.isNullOrUndefined(folder) || util.isNullOrUndefined(workspaceFolder.uri)) {
+      return this._folders.activeFolder?.folder;
     }
-    return this._folders.activeFolder?.folder;
+    return workspaceFolder;
   }
 
   private async _pickFolder() {
@@ -327,10 +378,12 @@ class ExtensionManager implements vscode.Disposable {
         // Do nothing. User cancelled
         return;
       }
-      const perist_message
-          = chosen.doConfigure ?
+      const perist_message = chosen.doConfigure ?
             localize('always.configure.on.open', 'Always configure projects upon opening?') :
-            localize('never.configure.on.open', 'Never configure projects on opening?');
+            localize('never.configure.on.open', 'Configure projects on opening?');
+      const button_messages = chosen.doConfigure ?
+            [ localize('yes.button', 'Yes'), localize('no.button', 'No') ] :
+            [ localize('never.button', 'Never'), localize('never.for.this.workspace.button', 'Not this workspace') ];
       interface Choice2 {
         title: string;
         persistMode: 'user'|'workspace';
@@ -341,8 +394,8 @@ class ExtensionManager implements vscode.Disposable {
                 .showInformationMessage<Choice2>(
                     perist_message,
                     {},
-                    {title: localize('yes.button', 'Yes'), persistMode: 'user'},
-                    {title: localize('for.this.workspace.button', 'For this Workspace'), persistMode: 'workspace'},
+                    {title: button_messages[0], persistMode: 'user'},
+                    {title: button_messages[1], persistMode: 'workspace'},
                     )
                 .then(async choice => {
                   if (!choice) {
@@ -386,7 +439,7 @@ class ExtensionManager implements vscode.Disposable {
       } else if (!ws) {
         // When adding a folder but the focus is on somewhere else
         // Do nothing but make sure we are showing the active folder correctly
-        this._statusBar.reloadVisibility();
+        this._statusBar.update();
       }
     }
   }
@@ -437,6 +490,7 @@ class ExtensionManager implements vscode.Disposable {
     this._folders.setActiveFolder(ws);
     this._statusBar.setActiveFolderName(ws?.name || '');
     this._statusBar.setActiveKitName(this._folders.activeFolder?.cmakeTools.activeKit?.name || '');
+    this._projectOutlineProvider.setActiveFolder(ws);
     this._setupSubscriptions();
   }
 
@@ -453,6 +507,7 @@ class ExtensionManager implements vscode.Disposable {
     }
   }
 
+  private cpptoolsNumFoldersReady: number = 0;
   private _updateCodeModel(folder: CMakeToolsFolder) {
     const cmt = folder.cmakeTools;
     this._projectOutlineProvider.updateCodeModel(
@@ -465,7 +520,7 @@ class ExtensionManager implements vscode.Disposable {
     );
     rollbar.invokeAsync(localize('update.code.model.for.cpptools', 'Update code model for cpptools'), {}, async () => {
       if (!this._cppToolsAPI) {
-        this._cppToolsAPI = await cpt.getCppToolsApi(cpt.Version.v2);
+        this._cppToolsAPI = await cpt.getCppToolsApi(cpt.Version.v4);
       }
       if (this._cppToolsAPI && cmt.codeModel && cmt.activeKit) {
         const codeModel = cmt.codeModel;
@@ -482,10 +537,14 @@ class ExtensionManager implements vscode.Disposable {
         const opts = drv ? drv.expansionOptions : undefined;
         const env = await effectiveKitEnvironment(kit, opts);
         const clCompilerPath = await findCLCompilerPath(env);
-        this._configProvider.updateConfigurationData({cache, codeModel, clCompilerPath});
+        this._configProvider.cpptoolsVersion = cpptools.getVersion();
+        this._configProvider.updateConfigurationData({cache, codeModel, clCompilerPath, activeTarget: cmt.defaultBuildTarget, folder: cmt.folder.uri.fsPath});
         await this.ensureCppToolsProviderRegistered();
-        if (cpptools.notifyReady) {
-          cpptools.notifyReady(this._configProvider);
+        if (cpptools.notifyReady && this.cpptoolsNumFoldersReady < this._folders.size) {
+          ++this.cpptoolsNumFoldersReady;
+          if (this.cpptoolsNumFoldersReady === this._folders.size) {
+            cpptools.notifyReady(this._configProvider);
+          }
         } else {
           cpptools.didChangeCustomConfiguration(this._configProvider);
         }
@@ -511,14 +570,14 @@ class ExtensionManager implements vscode.Disposable {
       this._statusBar.setVisible(true);
       this._statusMessageSub = cmt.onStatusMessageChanged(FireNow, s => this._statusBar.setStatusMessage(s));
       this._targetNameSub = cmt.onTargetNameChanged(FireNow, t => {
-        this._statusBar.targetName = t;
+        this._statusBar.setBuildTargetName(t);
       });
       this._buildTypeSub = cmt.onBuildTypeChanged(FireNow, bt => this._statusBar.setBuildTypeLabel(bt));
       this._launchTargetSub = cmt.onLaunchTargetNameChanged(FireNow, t => {
         this._statusBar.setLaunchTargetName(t || '');
       });
-      this._ctestEnabledSub = cmt.onCTestEnabledChanged(FireNow, e => this._statusBar.ctestEnabled = e);
-      this._testResultsSub = cmt.onTestResultsChanged(FireNow, r => this._statusBar.testResults = r);
+      this._ctestEnabledSub = cmt.onCTestEnabledChanged(FireNow, e => this._statusBar.setCTestEnabled(e));
+      this._testResultsSub = cmt.onTestResultsChanged(FireNow, r => this._statusBar.setTestResults(r));
       this._isBusySub = cmt.onIsBusyChanged(FireNow, b => this._statusBar.setIsBusy(b));
       this._statusBar.setActiveKitName(cmt.activeKit ? cmt.activeKit.name : '');
     }
@@ -806,6 +865,11 @@ class ExtensionManager implements vscode.Disposable {
     vscode.window.showErrorMessage(localize('compilation information.not.found', 'Unable to find compilation information for this file'));
   }
 
+  async selectWorkspace(folder?: vscode.WorkspaceFolder) {
+    if (!folder) return;
+    await this._setActiveFolder(folder);
+  }
+
   ctest(folder?: vscode.WorkspaceFolder) { return this.mapCMakeToolsFolder(cmt => cmt.ctest(), folder); }
 
   ctestAll() { return this.mapCMakeToolsAll(cmt => cmt.ctest()); }
@@ -865,6 +929,7 @@ class ExtensionManager implements vscode.Disposable {
 
   async hideLaunchCommand(shouldHide: boolean = true) {
     // Don't hide command selectLaunchTarget here since the target can still be useful, one example is ${command:cmake.launchTargetPath} in launch.json
+    this._statusBar.hideLaunchButton(shouldHide);
     await util.setContextValue(HIDE_LAUNCH_COMMAND_KEY, shouldHide);
   }
 
@@ -883,9 +948,12 @@ let _EXT_MANAGER: ExtensionManager|null = null;
 
 async function setup(context: vscode.ExtensionContext, progress: ProgressHandle) {
   reportProgress(progress, localize('initial.setup', 'Initial setup'));
-  await util.setContextValue('cmakeToolsActive', true);
+
   // Load a new extension manager
   const ext = _EXT_MANAGER = await ExtensionManager.create(context);
+
+  // Enable full or partial feature set, depending on state variable.
+  ext.enableFullFeatureSet(!ext.getActiveFolderContext()?.ignoreCMakeListsMissing);
 
   // A register function that helps us bind the commands to the extension
   function register<K extends keyof ExtensionManager>(name: K) {
@@ -960,6 +1028,7 @@ async function setup(context: vscode.ExtensionContext, progress: ProgressHandle)
     'resetState',
     'viewLog',
     'compileFile',
+    'selectWorkspace',
     'tasksBuildCommand',
     'hideLaunchCommand',
     'hideDebugCommand'
@@ -1005,6 +1074,8 @@ async function setup(context: vscode.ExtensionContext, progress: ProgressHandle)
                                       (what: TargetNode) => what.openInCMakeLists()),
       vscode.commands.registerCommand('cmake.outline.compileFile',
                                       (what: SourceFileNode) => runCommand('compileFile', what.filePath)),
+      vscode.commands.registerCommand('cmake.outline.selectWorkspace',
+                                      (what: WorkspaceFolderNode) => runCommand('selectWorkspace', what.wsFolder)),
   ]);
 }
 
@@ -1036,6 +1107,9 @@ export async function activate(context: vscode.ExtensionContext) {
         await vscode.window.showWarningMessage(localize('uninstall.old.cmaketools', 'Please uninstall any older versions of the CMake Tools extension. It is now published by Microsoft starting with version 1.2.0.'));
     }
 
+    // Silent re-scan when detecting a breaking change in the kits definition.
+    await scanForKitsIfNeeded(context);
+
   // Register a protocol handler to serve localized schemas
   vscode.workspace.registerTextDocumentContentProvider('cmake-tools-schema', new SchemaProvider());
   vscode.commands.executeCommand("setContext", "inCMakeProject", true);
@@ -1051,6 +1125,10 @@ export async function activate(context: vscode.ExtensionContext) {
 
   // TODO: Return the extension API
   // context.subscriptions.push(vscode.commands.registerCommand('cmake._extensionInstance', () => cmt));
+}
+
+export async function enableFullFeatureSet(fullFeatureSet: boolean) {
+    _EXT_MANAGER?.enableFullFeatureSet(fullFeatureSet);
 }
 
 // this method is called when your extension is deactivated