+++ src/tree.ts
@@ -135,7 +135,7 @@ function sortStringForType(type: codemodel_api.TargetTypeString): string {
 
 export class DirectoryNode<Node extends BaseNode> extends BaseNode {
   constructor(readonly prefix: string, readonly parent: string, readonly pathPart: string) {
-    super(`${prefix}::${path.join(parent, pathPart)}`);
+    super(`${prefix}${path.sep}${path.normalize(pathPart)}`);
   }
 
   private _subdirs = new Map<string, DirectoryNode<Node>>();
@@ -210,7 +210,10 @@ export class DirectoryNode<Node extends BaseNode> extends BaseNode {
 }
 
 export class SourceFileNode extends BaseNode {
-  constructor(readonly targetName: string, readonly filePath: string) { super(`${targetName}::${filePath}`); }
+  constructor(readonly prefix: string, readonly sourcePath: string, readonly filePath: string, private readonly _language?: string) {
+    // id: {prefix}::filename:directory of file relative to Target
+    super(`${prefix}::${path.basename(filePath)}:${path.relative(sourcePath, path.dirname(filePath))}`);
+  }
 
   get name() { return path.basename(this.filePath); }
 
@@ -222,7 +225,10 @@ export class SourceFileNode extends BaseNode {
     const item = new vscode.TreeItem(path.basename(this.filePath));
     item.id = this.id;
     item.resourceUri = vscode.Uri.file(this.filePath);
-    item.contextValue = 'nodeType=file';
+    const name = this.name.toLowerCase();
+    const cml = name == "cmakelists.txt";
+    const is_compilable = ['CXX', 'C'].indexOf(this._language||'')!==-1;
+    item.contextValue = ['nodeType=file', `compilable=${is_compilable}`, `cmakelists=${cml}`].join(',');
     item.command = {
       title: localize('open.file', 'Open file'),
       command: 'vscode.open',
@@ -233,8 +239,9 @@ export class SourceFileNode extends BaseNode {
 }
 
 export class TargetNode extends BaseNode {
-  constructor(readonly projectName: string, cm: codemodel_api.CodeModelTarget, readonly folder: vscode.WorkspaceFolder) {
-    super(`${projectName}::${cm.name}`);
+  constructor(readonly prefix: string, readonly projectName: string, cm: codemodel_api.CodeModelTarget, readonly folder: vscode.WorkspaceFolder) {
+    // id: {prefix}::target_name:target_path
+    super(`${prefix}::${cm.fullName || cm.name}:${cm.sourceDirectory || ''}`);
     this.name = cm.name;
     this.sourceDir = cm.sourceDirectory || '';
     this._rootDir = new DirectoryNode<SourceFileNode>(this.id, this.sourceDir, '');
@@ -346,11 +353,11 @@ export class TargetNode extends BaseNode {
         }
         const src_dir = path.dirname(src);
         const relpath = path.relative(this.sourceDir, src_dir);
-        addToTree(tree, relpath, new SourceFileNode(this.name, src));
+        addToTree(tree, relpath, new SourceFileNode(this.id, this.sourceDir, src, grp.language));
       }
     }
 
-    addToTree(tree, '', new SourceFileNode(this.name, path.join(this.sourceDir, 'CMakeLists.txt')));
+    addToTree(tree, '', new SourceFileNode(this.id, this.sourceDir, path.join(this.sourceDir, 'CMakeLists.txt')));
 
     collapseTreeInplace(tree);
 
@@ -378,9 +385,12 @@ export class TargetNode extends BaseNode {
 }
 
 class ProjectNode extends BaseNode {
-  constructor(readonly name: string, readonly folder: vscode.WorkspaceFolder) { super(name); }
+  constructor(readonly name: string, readonly folder: vscode.WorkspaceFolder, readonly sourceDirectory: string) {
+    // id: project_name:project_directory
+    super(`${name}:${sourceDirectory}`);
+   }
 
-  private readonly _rootDir = new DirectoryNode<TargetNode>('', '', '');
+  private readonly _rootDir = new DirectoryNode<TargetNode>(this.id, '', '');
 
   getOrderTuple() { return []; }
 
@@ -391,6 +401,7 @@ class ProjectNode extends BaseNode {
     if (this.getChildren().length === 0) {
       item.label += ` — (${localize('empty.project', 'Empty project')})`;
     }
+    item.tooltip=`${this.name}\n${this.sourceDirectory}`;
     return item;
   }
 
@@ -417,7 +428,7 @@ class ProjectNode extends BaseNode {
       context: ctx,
       update: (tgt, cm) => tgt.update(cm, ctx),
       create: cm => {
-        const node = new TargetNode(this.name, cm, this.folder);
+        const node = new TargetNode(this.id, this.name, cm, this.folder);
         node.update(cm, ctx);
         return node;
       },
@@ -433,16 +444,29 @@ interface ExternalUpdateContext {
   defaultTarget?: string;
 }
 
-class WorkspaceFolderNode extends BaseNode {
+export class WorkspaceFolderNode extends BaseNode {
   constructor(readonly wsFolder: vscode.WorkspaceFolder) { super(`wsf/${wsFolder.uri.fsPath}`); }
   private _children: BaseNode[] = [];
+
+  private _active: boolean = false;
+  setActive(active:boolean) {
+    this._active = active;
+  }
 
   getOrderTuple() { return [this.id]; }
 
   getTreeItem() {
     const item = new vscode.TreeItem(this.wsFolder.uri.fsPath, vscode.TreeItemCollapsibleState.Expanded);
     item.iconPath = vscode.ThemeIcon.Folder;
-    item.label += ' — Workspace Folder';
+    item.id = this.wsFolder.uri.fsPath;
+    let description:string;
+    if (this._active) {
+      description = localize('workspace.active', 'Active Workspace');
+    } else {
+      description = localize('workspace', 'Workspace');
+    }
+    item.description = `[${description}]`;
+    item.contextValue = ['nodeType=workspace', `selected=${this._active}`].join(',');
     return item;
   }
 
@@ -458,7 +482,7 @@ class WorkspaceFolderNode extends BaseNode {
     const config = model.configurations[0];
     const new_children: BaseNode[] = [];
     for (const pr of config.projects) {
-      const item = new ProjectNode(pr.name, ctx.folder);
+      const item = new ProjectNode(pr.name, ctx.folder, pr.sourceDirectory);
       item.update(pr, ctx);
       new_children.push(item);
     }
@@ -475,6 +499,7 @@ export class ProjectOutlineProvider implements vscode.TreeDataProvider<BaseNode>
   get onDidChangeTreeData() { return this._changeEvent.event; }
 
   private readonly _folders = new Map<string, WorkspaceFolderNode>();
+  private _selected_workspace?:WorkspaceFolderNode;
 
   addAllCurrentFolders() {
     for (const wsf of vscode.workspace.workspaceFolders || []) {
@@ -495,7 +520,7 @@ export class ProjectOutlineProvider implements vscode.TreeDataProvider<BaseNode>
   updateCodeModel(folder: vscode.WorkspaceFolder, model: codemodel_api.CodeModelContent|null, ctx: ExternalUpdateContext) {
     let existing = this._folders.get(folder.uri.fsPath);
     if (!existing) {
-      rollbar.error(localize('error.update.code.model.on.nonexist.folder', 'Updating code model on folder that does not yet exist?'));
+      rollbar.error(localize('error.update.code.model.on.nonexist.folder', 'Updating code model on folder that does not yet exist.'));
       // That's an error, but we can keep going otherwise.
       existing = new WorkspaceFolderNode(folder);
       this._folders.set(folder.uri.fsPath, existing);
@@ -505,9 +530,6 @@ export class ProjectOutlineProvider implements vscode.TreeDataProvider<BaseNode>
     existing.updateCodeModel(model, {...ctx, nodesToUpdate: updates, folder});
 
     this._changeEvent.fire(null);
-    for (const node of updates) {
-      this._changeEvent.fire(node);
-    }
   }
 
   getChildren(node?: BaseNode): BaseNode[] {
@@ -526,6 +548,21 @@ export class ProjectOutlineProvider implements vscode.TreeDataProvider<BaseNode>
       rollbar.error(localize('error.rendering.children.nodes', 'Error while rendering children nodes'));
       return [];
     }
+  }
+
+  setActiveFolder(ws: vscode.WorkspaceFolder | undefined):void {
+    if (!ws) return;
+    const current_node = this._selected_workspace;
+    const new_node = this._folders.get(ws.uri.fsPath);
+    if (current_node) {
+      current_node.setActive(false);
+      this._changeEvent.fire(current_node);
+    }
+    if (new_node) {
+      new_node.setActive(true);
+      this._changeEvent.fire(new_node);
+    }
+    this._selected_workspace = new_node;
   }
 
   async getTreeItem(node: BaseNode) { return node.getTreeItem(); }